# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: Kizashi API
  version: "1.0.0"
  description: |
    Kizashi REST API — エンドポイント検知・対応プラットフォーム

    ## 認証
    すべての保護されたエンドポイントは `Authorization: Bearer <JWT>` ヘッダーが必要です。
    `/api/v1/auth/login` でJWTトークンを取得してください。

    ## レート制限
    - 認証済みリクエスト: 1000 req/分
    - 認証エンドポイント: 10 req/分
  contact:
    name: Kizashi Support
    email: support@kizashi-edr.example.com
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080
    description: 開発サーバー
  - url: https://api.kizashi-edr.example.com
    description: 本番サーバー
security:
  - BearerAuth: []
...
paths:
...
  /api/v1/reports:
    get:
      operationId: listReports
      tags: [reports]
      summary: レポート一覧取得
      responses:
        "200":
          description: レポート一覧
        "402":
          $ref: "#/components/responses/PaymentRequired"
    post:
      operationId: createReport
      tags: [reports]
      summary: レポート生成
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type]
              properties:
                type:
                  type: string
                  enum: [executive, compliance, incident, threat]
                period_days:
                  type: integer
                  default: 30
      responses:
        "201":
          description: レポートを生成しました
        "402":
          $ref: "#/components/responses/PaymentRequired"

  # ── System ─────────────────────────────────────────────────────
...
  /api/v1/reports/{id}:
    get:
      operationId: getReport
      tags: [reports]
      summary: レポート詳細取得
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: レポート詳細
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Report" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "402": { $ref: "#/components/responses/PaymentRequired" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      operationId: deleteReport
      tags: [reports]
      summary: レポート削除
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: 削除成功
        "401": { $ref: "#/components/responses/Unauthorized" }
        "402": { $ref: "#/components/responses/PaymentRequired" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ID:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
...
    Unauthorized:
      description: 認証が必要です
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    NotFound:
      description: リソースが見つかりません
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    PaymentRequired:
      description: |
        現在のライセンスプランで利用可能な機能・上限を超えています。
        レスポンスボディの `error` フィールドで原因を判別:
        - `plan_required`: 機能ゲート (例: MDM を Lite テナントが叩いた) — `feature` / `required_plan` / `upgrade_message` を含む
        - `agent_limit_reached`: エージェント上限超過 — `agent_limit` / `agents_active` / `plan` を含む
      content:
        application/json:
          schema:
            oneOf:
              - $ref: "#/components/schemas/PlanRequiredError"
              - $ref: "#/components/schemas/AgentLimitReachedError"
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: エラーメッセージ
    PlanRequiredError:
      type: object
      description: |
        `RequireFeature` ミドルウェアが返す 402 エラー (`server/internal/middleware/plan_gate.go`)。
        現在のライセンスが必要な機能を含まない場合に返される。
      required: [error, feature, required_plan, upgrade_message]
      properties:
        error:
          type: string
          enum: [plan_required]
        feature:
          type: string
          description: |
            アクセスを試みた機能の名前 (`server/internal/license/manager.go` の `Feature*` 定数)。
          enum:
            - basic_detection
            - alerts
            - reports
            - mdm
            - ai_investigation
            - siem_integration
            - playbooks
            - threat_intel
            - yara
            - ml_detection
            - threat_hunting
            - multi_tenant
            - compliance
            - api_access
            - xdr
            - deception
            - forensics
            - soar
        required_plan:
          type: string
          description: |
            この機能を解除する最小プラン。`requiredPlanFor` ヘルパーが返す値で、
            v1.3.3 で MDM 機能は `enterprise` から `starter` に修正済み。
            `lite` は他機能を持たない (Lite が含む機能は 200 を返すため `required_plan` は出現しない)。
          enum: [starter, professional, enterprise]
        upgrade_message:
          type: string
          example: この機能はStarterプラン（¥1,800/端末/月）以上で利用できます
    AgentLimitReachedError:
      type: object
      description: |
        `EnforceAgentLimit` ミドルウェアが返す 402 エラー (`server/internal/middleware/plan_gate.go`)。
        プランの `agent_limit` に達した状態でエージェント追加を試みると返される。
        Lite プランでは Stripe Subscription Quantity が `agent_limit` になる
        (例: 10 EP 契約のテナントは 10 台で上限)。
      required: [error, agent_limit, agents_active, plan, upgrade_message]
      properties:
        error:
          type: string
          enum: [agent_limit_reached]
        agent_limit:
          type: integer
          description: 現プランのエージェント上限 (0=無制限)
          example: 10
        agents_active:
          type: integer
          description: 現在アクティブなエージェント数
          example: 10
        plan:
          type: string
          enum: [free, lite, starter, professional, enterprise]
        upgrade_message:
          type: string
          example: エージェント数がプランの上限に達しています。プランをアップグレードしてください。
...
    Report:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [executive, compliance, incident, threat]
        status:
          type: string
          enum: [pending, generating, completed, failed]
        created_by:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        download_url:
          type: string
...